==> study4/source/01_preregistered-analyses.R <==
# load packages -----------------------------------------------------------

if(!require('pacman')) {
  install.packages('pacman')
}
pacman::p_load(tidyverse, here)

# load data ---------------------------------------------------------------

clean <- read_csv(here("study4", "data", "clean.csv"))


## recoding to factors for vars used in logits

clean$gender <- factor(clean$gender)
clean$comp_choice <- factor(clean$comp_choice)
clean$condition <- factor(clean$condition)

## fraud only version of dataset

clean_fraud_removed <- clean %>% filter(fraud == 0) 


## creating practice cond & women only versions of dataset

practice <- clean_fraud_removed %>% filter(condition == "pract")
women <- clean_fraud_removed %>% filter(gender == "Woman")

# primary hypothesis 1 ----------------------------------------------------

## decided against using z test to test main hypo because it is less commonly used & to be able to easily compare effects while adding controls to other models

# prep_comp <- nrow(clean %>% filter(gender == "Woman" & condition == "pract" & comp_choice == "tournament"))
# total_prep<- nrow(clean %>% filter(gender == "Woman" & condition == "pract"))
# 
# con_comp <- nrow(clean %>% filter(gender == "Woman" & condition == "control" & comp_choice == "tournament"))
# total_con <- nrow(clean %>% filter(gender == "Woman" & condition == "control"))
# 
# 
# ## https://www.r-bloggers.com/comparison-of-two-proportions-parametric-z-test-and-non-parametric-chi-squared-methods/
# 
# z.prop = function(x1,x2,n1,n2){
#   numerator = (x1/n1) - (x2/n2)
#   p.common = (x1+x2) / (n1+n2)
#   denominator = sqrt(p.common * (1-p.common) * (1/n1 + 1/n2))
#   z.prop.ris = numerator / denominator
#   return(z.prop.ris)
# }
# 
# hypothesis1_z <- z.prop(prep_comp, con_comp, total_prep, total_con)
# hypothesis1_pval <- pnorm(hypothesis1_z)  

hypothesis1  <- glm(comp_choice~condition, family = binomial, data = women)


# primary hypothesis 2 ----------------------------------------------------

hypothesis2a <- glm(extra_practice_rounds_count ~ gender,family="poisson",data = practice)

hypothesis2b <- glm(extra_practice_rounds_count ~ gender + task_score,family="poisson",data = practice)

hypothesis2c <- glm(extra_practice_rounds_count ~ gender + task_score + risk + conf_rank,family="poisson",data = practice)

# primary hypothesis 3 ----------------------------------------------------

hypothesis3a <- glm(comp_choice ~ gender*condition,family=binomial,data = clean_fraud_removed)

hypothesis3b <- glm(comp_choice ~ gender*condition + task_score,family=binomial,data = clean_fraud_removed)

hypothesis3c <- glm(comp_choice ~ gender*condition + task_score + risk + conf_rank,family=binomial,data = clean_fraud_removed)
